A further large set of small changes. Mostly signedness and constness issues, and lots of inlining in class Mineserver.
The type for "map" is now size_t, not int, as it is an index of a std::vector.
Made a few comments at places where a line of code looks questionable.

+++ src/mob.h
@@ -45,7 +45,7 @@ public:
   uint32_t UID;
   int8_t type;
   double x, y, z;
-  int map;
+  size_t map;
   int8_t yaw, pitch;
   int8_t meta;
   bool spawned;
